+++ 02hw/search-engine/cmd/app/gosearch.go
@@ -3,7 +3,7 @@ package main
 import (
 	"flag"
 	"fmt"
-	"go-core-4/02hw/search-engine/pkg/scanner"
+	"go-core-4/02hw/search-engine/pkg/crawler/spider"
 	"strings"
 )
 
@@ -16,14 +16,14 @@ var s = flag.String("s", "", "search argument")
 
 func main() {
 	flag.Parse()
-	urls := scanner.URLs([]string{godev, practicalgo})
+	urls := scanResult([]string{godev, practicalgo})
 
 	for _, v := range search(*s, urls) {
 		fmt.Println(v)
 	}
 }
 
-// ищет строку в слайсе урлов
+// поиск строки в слайсе урлов
 func search(s string, urls []string) []string {
 	var res []string
 
@@ -34,4 +34,30 @@ func search(s string, urls []string) []string {
 	}
 
 	return res
+}
+
+func scanResult(urls []string) []string {
+	var data []string
+
+	for _, url := range urls {
+		urls, err := scan(url)
+		if err != nil {
+			continue
+		}
+		data = append(data, urls...)
+	}
+	return data
+}
+
+func scan(url string) (urls []string, err error) {
+	s := spider.New()
+	var res []string
+	docs, err := s.Scan(url, 2)
+	if err != nil {
+		return nil, err
+	}
+	for _, doc := range docs {
+		res = append(res, doc.URL)
+	}
+	return res, nil
 }
